refactor location modal
